Response Code Structure Improvement, updated gitignore
New custom Exception made to handle response code cases. However, this
specific one (UnexpectedResponseCodeException) isn't expected to be used
in practice, but is more for the sake of completeness (and compilation).
Furthermore, another file was included in gitignore as I'm using JUnit
testing to check the classes rather than multiple main methods.